+++ src/main/java/com/ecomarket/cl/ecomarket/repository/ResenasRepository.java
@@ -3,5 +3,5 @@ package com.ecomarket.cl.ecomarket.repository;
 import com.ecomarket.cl.ecomarket.model.Resenas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResenasRepository extends JpaRepository<Resenas, Long> {
+public interface ResenasRepository extends JpaRepository<Resenas, Integer> {
 }
